# Improve decoding performance