--- a/src/cjs/index.cjs
+++ b/src/cjs/index.cjs
@@ -81,7 +81,7 @@ function base (ALPHABET) {
     const size = (((source.length - psz) * FACTOR) + 1) >>> 0 // log(58) / log(256), rounded up.
     const b256 = new Uint8Array(size)
     // Process the characters.
-    while (source[psz]) {
+    while (psz < source.length) {
       // Decode character
       let carry = BASE_MAP[source.charCodeAt(psz)]
       // Invalid character