@steveluscher 2024-07-04
You don’t feel like backporting this to bs58 version 4.x, do you @junderw? There are reasons we can’t upgrade to v6 and have to stay on 4.x.