@junderw 2024-07-04
Give me a list of commands to backport and publish all these backports without updating the npm latest tag for everything, and I'll consider running those commands when I have free time.

It would probably take me a few minutes to find everything, cherry-pick, push a branch for previous major version(s) push a tag to Github, publish to npm under a new non-latest tag etc. And then I would need to do the same for bs58 and bump the dependency etc.

Probably about 15-20 minutes of work depending on how much I can concentrate without messing up.

But if you give me a list of commands and I can check that they're sane, running them would be a few seconds.

Assuming base-x and bs58 are sibling directories and these repos are `origin` remotes is fine.