@steveluscher 2024-07-04
For sure!

```
# Rewind to base-x@3.0.9
cd base-x
git checkout v3.0.9
git clean -dfx .
npm i

# Backport change, build, and test
git cherry-pick 261fb3636147815a7705375cfe7c1217db15e8f5
rm -rf src
npm run build
git add -A
git cherry-pick --continue --no-edit
npm run test

# Publish base-x@3.0.10
npm version 3.0.10
npm publish

# Create a backport branch for posterity and push the whole thing to Github
git checkout -b v3.x
git push origin v3.x
```

That's it. `bs58` will pick up this new version by virtue of it depending on `^3.0.2`.